requirements.txt
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
dotenv==0.9.9
Flask==3.1.2
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
pandas==2.3.2
PyJWT==2.10.1
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
xlsxwriter==3.2.9
